requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
drf-nested-routers>=0.95.0
python-docx>=1.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
annotated-types==0.7.0
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
lxml==6.0.2
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-docx==1.2.0
python-dotenv==1.2.1
pdfminer.six>=20221105
PyMuPDF>=1.24.0
pypdf>=4.2.0
cryptography>=42.0.0
xhtml2pdf>=0.2.15
sqlparse==0.5.5
gdown>=5.0
pytesseract>=0.3.10
python-pptx>=0.6.23
typing-inspection==0.4.2
typing_extensions==4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
openpyxl>=3.1.0
xlrd>=2.0.0
celery>=5.4.0
redis>=5.0.0
